==> Indian_Markets/Equity_Mutual_Funds_v1/notebooks/Step 1.1 - Download historical benchmarks data.ipynb <==
#### Use this Code to download historical data for Indian Market benchmarks

import yfinance as yf

# Download Nifty 50 data
df = yf.download(['^NSEI', '^BSESN', '^CRSLDX','^NSEMDCP50'],  start='2016-12-31', end='2024-05-01')[['Adj Close']]


Index	Description	Yahoo Finance Ticker

Nifty 50	Top 50 companies on NSE	^NSEI

BSE Sensex	Top 30 companies on BSE	^BSESN

Nifty Midcap 50	Top 50 mid-cap companies on NSE	^NSEMDCP50

Nifty 500 ^CRSLDX


df.to_csv(r'../data/market_benchmarks_historical_data/benchmark_data.csv')

df

# !pip install yfinance
