==> Indian_Markets/Equity_Mutual_Funds_v1/notebooks/Step 2.1 - Generate Market Scenarios - Historical.ipynb <==
import pandas as pd
import numpy as np

data_dir = r'../data/mutual_funds_data'
all_mf_ts_fname = 'selected_mutualfunds_nav_timeseries_data.csv'
all_mf_ts_df = pd.read_csv(data_dir+'/'+all_mf_ts_fname)
all_mf_ts_df['Date'] = pd.to_datetime(all_mf_ts_df['Date'])
sim_save_dir = r'../data/simulations/market_scenarios_historical'

all_mf_ts_df = all_mf_ts_df.drop('Unnamed: 0', axis=1)
all_mf_ts_df = all_mf_ts_df.set_index('Date')

all_mf_ts_df

all_mf_ts_df['Year'] = all_mf_ts_df.index.year
ts_grouped_by_year = all_mf_ts_df.groupby('Year')

for group_name, group_df in ts_grouped_by_year:
    get_ret_vol_corr_from_ts(group_df, name=group_name)


def get_ret_vol_corr_from_ts(tmpdf, name='tmp'):
    # Calculate for all together to maintain consistency of correlations etc.
    # This could impact accuracy to some extent as we might miss a few dates in the beginning or here and there.
    # As a sanity check, we see it only misses a few dates (less than 5) per year across the mfs
    name=str(name)
    tmpdf_cleaned = tmpdf[(tmpdf!=0).all(axis=1)]
    tmpdf_cleaned = tmpdf_cleaned.drop(columns='Year')
    annual_returns = tmpdf_cleaned.apply(lambda x: (x.iloc[-1] / x.iloc[0]) -1)

    daily_returns = tmpdf_cleaned.pct_change()
    daily_returns = daily_returns.dropna()
    annual_vol = daily_returns.apply(lambda x: x.std()*np.sqrt(252))

    annual_corr = daily_returns.corr()
    
    annual_returns.to_csv(sim_save_dir + '//'+name+'_annual_ret.csv', index=True, header=False)
    annual_vol.to_csv(sim_save_dir + '//'+name+'_annual_vol.csv', index=True, header=False)
    annual_corr.to_csv(sim_save_dir + '//'+name+'_annual_corr.csv', index=True, header=True)
    return True


    annual_returns.to_csv(sim_save_dir + '//'+name+'_annual_ret.csv', index=True, header=False)
    annual_vol.to_csv(sim_save_dir + '//'+name+'_annual_vol.csv', index=True, header=False)
    annual_corr.to_csv(sim_save_dir + '//'+name+'_annual_corr.csv', index=True, header=True)

# Run following cells manually for all years 2017-2024

name='2017'

tmpdf_cleaned = tmpdf[(tmpdf!=0).all(axis=1)]
tmpdf_cleaned = tmpdf_cleaned.drop(columns='Year')
annual_returns = tmpdf_cleaned.apply(lambda x: (x.iloc[-1] / x.iloc[0]) -1)
annual_returns
daily_returns = tmpdf_cleaned.pct_change()
daily_returns = daily_returns.dropna()
annual_vol = daily_returns.apply(lambda x: x.std()*np.sqrt(252))
annual_vol
annual_corr = daily_returns.corr()
annual_corr

tmpdf_cleaned.to_clipboard()

annual_vol
