==> Indian_Markets/Equity_Mutual_Funds_v1/notebooks/Step 3.1 - Generate_Portfolio_runs on Historical Mkt Data.ipynb <==
import pandas as pd
import numpy as np
import csv
import os
from tqdm import tqdm

sim_save_dir = r'../data/simulations/market_scenarios_historical'
portfolio_run_out_dir = r'../data/simulations/portfolio_simulations_historical'


def gen_pf_wts(total_n, n, num_pf_wt_generations):
    final_list = []
    for i in range(num_pf_wt_generations):
        selected_positions = np.random.choice(range(total_n), size = n, replace=False)
        random_weights = np.random.rand(n)
        random_weights /= np.sum(random_weights)
        random_weights = np.round(random_weights,2)
        portfolio_weights = np.zeros(total_n)
        portfolio_weights[selected_positions] = random_weights
        final_list.append(portfolio_weights)

    return final_list

def generate_portfolio_ret_vol(pf_wts, ret_np, vol_np, corr_np):
    # Assuming every input is np.array
    cov_matrix = np.outer(vol_np, vol_np) * corr_np
    
    portfolio_return = round(np.dot(pf_wts, ret_np),3)
    portfolio_vol = round(np.sqrt(np.dot(pf_wts.T, np.dot(cov_matrix, pf_wts))),3)

    return portfolio_return, portfolio_vol

# test gen_portfolio_ret_vol
pf_wts = np.array([0.3, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.7, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])
t_year = '2020'
t_ret = pd.read_csv(sim_save_dir+'//'+t_year+'_annual_ret.csv', header=None)
t_ret.columns = ['mf_id','annual_return']
t_vol = pd.read_csv(sim_save_dir+'//'+t_year+'_annual_vol.csv', header=None)
t_vol.columns = ['mf_id','annual_vol']
t_corr = pd.read_csv(sim_save_dir+'//'+t_year+'_annual_corr.csv', index_col=0)

portfolio_return, portfolio_vol = generate_portfolio_ret_vol(pf_wts, np.array(t_ret['annual_return']), np.array(t_vol['annual_vol']), np.array(t_corr))
print('pf_ret', portfolio_return) # Manually checked return
print('pf_vol', portfolio_vol)


t_corr.to_clipboard()

def write_results_to_csv(fname, year, pf_wts, portfolio_return, portfolio_vol):
    with open(os.path.join(portfolio_run_out_dir, fname, fname+'_'+year+'.csv'), mode='a', newline='') as f:
        writer = csv.writer(f)    
        writer.writerow(np.append(np.append(pf_wts,portfolio_return), portfolio_vol))
    return True

total_n = 25
k = 3
num_pf_wt_generations = 1000000
pf_wts_list = gen_pf_wts(total_n, k, num_pf_wt_generations)

for year in tqdm(range(2017,2025)):
    
    year = str(year)
    annual_ret = pd.read_csv(sim_save_dir+'//'+year+'_annual_ret.csv', header=None)
    annual_ret.columns = ['mf_id','annual_return']
    annual_vol = pd.read_csv(sim_save_dir+'//'+year+'_annual_vol.csv', header=None)
    annual_vol.columns = ['mf_id','annual_vol']
    annual_corr = pd.read_csv(sim_save_dir+'//'+year+'_annual_corr.csv', index_col=0)
    ret_np = np.array(annual_ret['annual_return'])
    vol_np = np.array(annual_vol['annual_vol'])
    corr_np = np.array(annual_corr)
    fname = 'k_'+str(k)

    for pf_wts in pf_wts_list:
        portfolio_return, portfolio_vol = generate_portfolio_ret_vol(pf_wts, ret_np, vol_np, corr_np)
        write_results_to_csv(fname, year,  pf_wts, portfolio_return, portfolio_vol)
