==> Indian_Markets/Equity_Mutual_Funds_v1/notebooks/Step 7 - Generate PnL time series from portfolio wts.ipynb <==
### Compile all results together, generate a time series of returns given portfolio weights from various strategies such as GOPALS, MVO, equal benchmarks etc.

# Generate yearly pnl time series given portfolio weights and historical timeseries of mutual funds
import os
import pandas as pd
import numpy as np


pf_wt_dir = r'../results/backtests/portfolio_weights'
hist_ts_file = r'../data/mutual_funds_data/selected_mutualfunds_nav_timeseries_data.csv'
ret_ts_dir = r'../results/backtests/strategy_returns_timeseries'


mkt_df = pd.read_csv(hist_ts_file, index_col='Date', parse_dates=['Date'])

strategy_name = 'GOPALS_Strategy2.3'
n = 3
pct_= 0.95
# year = 2023 # later do this for all four years

for year in range(2020, 2024):
    pf_wt = pd.read_csv(os.path.join(pf_wt_dir, strategy_name, 'GOPALS_n_'+str(n)+'_year_'+str(year)+'_pct_'+str(pct_)+'.csv'), header=None)
    pf_wt = pf_wt/pf_wt.values.sum()
    # Load mkt data
    filtered_data = mkt_df[mkt_df.index.year.isin([year])]
    filtered_data = filtered_data[(filtered_data != 0.0).all(axis=1)]
    filtered_data_ret = filtered_data.pct_change().dropna()
    ret_ts = filtered_data_ret.dot(pf_wt.values)
    ret_ts.to_csv(os.path.join(ret_ts_dir, strategy_name+'_n_'+str(n)+'_year_'+str(year)+'_pct_'+str(pct_)+'.csv'))

# MVO
strategy_name = 'MVO'
for year in range(2020, 2024):
    pf_wt = pd.read_csv(os.path.join(pf_wt_dir, strategy_name, 'MVO_year_'+str(year)+'.csv'), header=None)
    pf_wt = pf_wt/pf_wt.values.sum()
    # Load mkt data
    filtered_data = mkt_df[mkt_df.index.year.isin([year])]
    filtered_data = filtered_data[(filtered_data != 0.0).all(axis=1)]
    filtered_data_ret = filtered_data.pct_change().dropna()
    ret_ts = filtered_data_ret.dot(pf_wt.values)
    ret_ts.to_csv(os.path.join(ret_ts_dir, strategy_name+'_year_'+str(year)+'.csv'))

# MVO
strategy_name = 'MVO_new'
for year in range(2020, 2024):
    pf_wt = pd.read_csv(os.path.join(pf_wt_dir, strategy_name, 'MVO_new_year_'+str(year)+'.csv'), header=None)
    pf_wt = pf_wt/pf_wt.values.sum()
    # Load mkt data
    filtered_data = mkt_df[mkt_df.index.year.isin([year])]
    filtered_data = filtered_data[(filtered_data != 0.0).all(axis=1)]
    filtered_data_ret = filtered_data.pct_change().dropna()
    ret_ts = filtered_data_ret.dot(pf_wt.values)
    ret_ts.to_csv(os.path.join(ret_ts_dir, strategy_name+'_year_'+str(year)+'.csv'))

# Generate for benchmarks
strategy_name = 'mkt_benchmark'
mkt_df1 = pd.read_csv(os.path.join(r'../data/market_benchmarks_historical_data', 'benchmark_data.csv'), index_col='Date',parse_dates=['Date'])

for year in range(2020, 2024):
    
    filtered_data = mkt_df1[mkt_df1.index.year.isin([year])]
    filtered_data = filtered_data[(filtered_data != 0.0).all(axis=1)]
    filtered_data_ret = filtered_data.pct_change().dropna()
    ret_ts = filtered_data_ret.copy()
    ret_ts.to_csv(os.path.join(ret_ts_dir, strategy_name+'_year_'+str(year)+'.csv'))


# Generate for Equally Weighted Portfolio

strategy_name = 'Equal_wtd'
for year in range(2020, 2024):
    pf_wt = np.ones(25)/25
    
    # Load mkt data
    filtered_data = mkt_df[mkt_df.index.year.isin([year])]
    filtered_data = filtered_data[(filtered_data != 0.0).all(axis=1)]
    filtered_data_ret = filtered_data.pct_change().dropna()
    ret_ts = filtered_data_ret.dot(pf_wt)
    ret_ts.to_csv(os.path.join(ret_ts_dir, strategy_name+'_year_'+str(year)+'.csv'))
